fix(api): count the people a claimed key kept out of the run
A user whose every nudge had its dedupe key claimed before the run started
was filtered out before the grouping, so they reached no decision and landed
in no count except a row level tally. On an hourly cron that is most of the
people the job looks at. They are grouped like everybody else now, with an
empty queue and the reason that fits.

The heartbeat also moves into a `finally`. The selectors, the claimed key
lookup and the cadence read all happen before any per user handling can catch
anything, so a database blip in one of them still ended the hour with no row
at all, and a missing hour is documented as a dead cron.

The selectors move to their own module. The service was over its line budget
and the two halves answer different questions: who is due, and who gets it.
Every name callers already imported is re-exported, so no call site moves.
